25.08. 3

+++ 57_GameDataParser/Model/VideoGame.cs
@@ -4,5 +4,5 @@ public class VideoGame
 {
     public string Title { get; init; }
     public int ReleaseYear { get; init; }
-    public double Rating { get; init; }
+    public decimal Rating { get; init; }
 }
